Módulos 'fwp_utilities', 'fwp_analysis', 'fwp_string', 'fwp_classes'. Script 'SMeasure'. Cambios: 1 Funcionales (+) + 9 Sintácticos (+).
Cambios funcionales (+):
	ahora 'PIDController' tiene un log de un único elemento y tal vez un par más de correcciones.

Cambios sintácticos (+):
	elimino la vieja clase 'InstancesDic' de 'fwp_classes';
	la actual clase 'InstancesDic' ahora se llama correctamente 'InstancesDict';
	las clases 'TypedList', 'ObjectView', 'ObjectDict' de 'fwp_classes' ahora están en 'fwp_utilities';
	la clase 'append_data_to_string' de 'fwp_analysis' ahora está en 'fwp_utilities';
	'fwp_classes.NoNoneList' ahora está incluida en 'fwp_utilities.NotCertainTypeList';
	las funciones 'error_value', 'multimeter_error' de 'fwp_string' ahora están en 'fwp_analysis';
	la función 'append_data_to_string' de 'fwp_analysis' ahora está en 'fwp_utilities';
	actualizo módulos y script 'SMeasure' según estos cambios.

Cambios sintácticos (-):
	actualizo documentación inicial de 'fwp_analysis';
	cambio de orden algunas funciones de 'fwp_analysis'.
