+++ fwp_classes.py
@@ -5,38 +5,6 @@ This module holds classes developed by our group.
 @author: Usuario
 """
 
-#%%
-
-class TypedList(list):
-    
-    """A list that only appends a certain type"""    
-    
-    def __init__(self, Class):
-        self.instance = Class
-
-    def append(self, item):
-        if not isinstance(item, self.instance):
-            raise TypeError('Nop because {}'.format(self.instance))
-        super(TypedList, self).append(item)
-
-# Simple, but apparently not the best way...
-# https://stackoverflow.com/questions/3487434/overriding-append-method-after-inheriting-from-a-python-list
-
-class NoNoneList(list):
-    
-    """A list that doesn't append None elements"""
-    
-    def __init__(self, *iterable, show_exceptions=True):
-        super().__init__(iterable)
-        self.show_exceptions = show_exceptions
-    
-    def append(self, item):
-        if item is not None:
-            super().append(item)
-        else:
-            if self.show_exceptions:
-                raise TypeError("Can't append a None element")
-
 #%%       
 
 class DynamicList(list):
@@ -106,44 +74,7 @@ class DynamicDict(dict):
 
 #%%
 
-"""
-This all started because we'd like something like DAQ.ai.ch1.state = True
-where "ch1" is a name generated by a dict saved on DAQ.
-Then we thought it would be nice if we had DAQ.ai.state = True that could
-act on all of "ch" saved on DAQ's dict.
-
-This is some code found online that could work as inspiration.
-"""
-
-class ObjectView(object):
-    
-    def __init__(self, d):
-        
-        self.__dict__ = d
-
-class ObjectDict(dict):
-    
-    def __getattr__(self, name):
-        
-        if name in self:
-            return self[name]
-        else:
-            raise AttributeError("No such attribute: " + name)
-
-    def __setattr__(self, name, value):
-        
-        self[name] = value
-
-    def __delattr__(self, name):
-        
-        if name in self:
-            del self[name]
-        else:
-            raise AttributeError("No such attribute: " + name)
-
-#%%
-
-class InstancesDic:
+class InstancesDict:
     
     """Example of a class that holds a callable dictionary of instances.
 
@@ -155,9 +86,9 @@ class InstancesDic:
     >> instance_a, instance_b = MyClass(), MyClass(12)
     >> instance_c = ClassWithInstances(dict(a=instance_a,
                                             b=instance_b))
-    >> Z = InstancesDic({1: instance_a,
-                         2: instance_b,
-                         3: instance_c})
+    >> Z = InstancesDict({1: instance_a,
+                          2: instance_b,
+                          3: instance_c})
     >> Z(1)
     <__main__.MyClass at 0x2e5572a2dd8>
     >> Z(1).sub_prop
@@ -549,171 +480,6 @@ class DottableMultiWrapper():
             
         instances = dict(instances)
         self.__dict__.update(instances)
-
-
-## Old class
-#class ClassWithInstances:
-#    
-#    """Example of a class which allows dot calling instances.
-#    
-#    Examples
-#    --------
-#    >> class MyClass:
-#    
-#        def __init__(self, value=10):
-#            self.sub_prop = value
-#            self._prop = value
-#        
-#        @property
-#        def prop(self):
-#            return self._prop
-#        
-#        @prop.setter
-#        def prop(self, value):
-#            self._prop = value
-#            
-#        def sub_method(self, item):
-#            return item * self.sub_prop
-#        
-#        def method(self, item):
-#            return item * self.prop
-#        
-#    >> Z = ClassWithInstances(a=MyClass(), b=MyClass(2))
-#    >>
-#    >> # Let's check dot calling
-#    >> Z.a.sub_prop 
-#    10
-#    >> Z.a.sub_prop = 1
-#    >> Z.a.sub_prop
-#    1
-#    >>
-#    >> # Let's check dot calling all instances at once
-#    >> Z.sub_prop
-#    {'a': 1, 'b': 2}
-#    >> Z.all.sub_prop = 3
-#    >> Z.sub_prop
-#    {'a': 3, 'b': 3}
-#    >> Z.all.sub_prop = {'a': 1}
-#    >> Z.sub_prop
-#    {'a': 1, 'b': 3}
-#    >> 
-#    >> # This also works with methods
-#    >> Z.a.prop
-#    10
-#    >> Z.a.method(2)
-#    20
-#    >> Z.all.prop
-#    {'a': 10, 'b': 2}
-#    >> Z.all.method(2)
-#    {'a': 20, 'b': 4}
-#    >>
-#    >> # This is an updatable class too
-#    >> Z.add(c=MyClass(4))
-#    >> Z.sub_prop
-#    {'a': 1, 'b': 3, 'c': 4}
-#    
-#    
-#    Warnings
-#    --------
-#    >> # There are some wrong ways of calling things
-#    >> Z.all.method()
-#    {'a': 20, 'b': 4}
-#    >> Z.method()
-#    RecursionError: maximum recursion depth exceeded
-#    >> Z.all.prop
-#    {'a': 10, 'b': 2}
-#    >> Z.prop
-#    RecursionError: maximum recursion depth exceeded
-#    
-#    """
-#    
-#    def __init__(self, **instances):
-#        
-#        self.__instances = {}
-#        self.__properties = {}
-#        self.all = WrapperDict()
-#        self.add(**instances)
-#    
-#    @property
-#    def instances(self):
-#        return self.__instances
-#    
-#    @instances.setter
-#    def instances(self, value):
-#        return AttributeError("See 'add' method instead.")
-#    
-#    @property
-#    def properties(self):
-#        return self.__properties
-#    
-#    @properties.setter
-#    def properties(self, value):
-#        return AttributeError("Shouldn't set this manually!")
-#    
-#    def add(self, **instances):
-#        
-#        instances = dict(instances)
-#        self.__dict__.update(instances)
-#        self.instances.update(instances)
-#        for name, ins in instances.items():
-#            self.properties.update({name: m for m in dir(ins)})
-#        self.all.update(**instances)
-#    
-#    def __getattr__(self, name):
-#        
-#        name = name.split('.')
-#        
-#        if len(name) == 1:
-#            
-#            name = name[0]
-#            
-#            if name in self.properties.values():
-#
-#                results = {}
-#                for n, ins in self.instances.items():
-#                    try:
-#                        results.update({n : #n + '_' + name : 
-#                            eval("ins.{}".format(name))})
-#                    except:
-#                        pass                
-#                return results
-#                
-#            else:
-#            
-#                return eval('self.{}'.format(name))
-#                
-#        else:
-#            
-#            instance_name = name[0]        
-#            attribute_name = name[-1]
-#            
-#            if instance_name not in self.__dict__.keys():
-#                raise AttributeError("No such attribute: " + name)
-#            
-#            command = "self.__dict__['{}']".format(instance_name)
-#            command = command + '.' + attribute_name
-#            
-#            return eval(command)
-#
-#    def __setattr__(self, name, value):
-#        
-#        name = name.split('.')
-#        
-#        if len(name) == 1:
-#            
-#            name = name[0]
-#            self.__dict__.update({name : value})
-#        
-#        else:
-#        
-#            class_name = name[0]        
-#            attribute_name = name[-1]
-#            
-#            command = "self.__dict__['{}']".format(class_name)
-#            command = command + '.' + attribute_name
-#            command = command + '=' + str(value)
-#            
-#            eval(command)
     
 #%%
 
@@ -889,7 +655,3 @@ class MultiWrapper(BigClass):
             
         instances = dict(instances)
         self.__dict__.update(instances)
-        
-
-#%%
-
